requirements.txt
callbacks==0.3.0
ghapi==1.0.6
jsonlines==4.0.0
midigpt==0.0.1b5
mido==1.3.3
nox==2025.5.1
numpy==2.2.5
pyaudio==0.2.14
pyfluidsynth==1.3.4
PyQt6==6.9.0
PyQt6_sip==13.10.0
pytest==8.3.5
rich==14.0.0
setuptools==78.1.0
tqdm==4.67.1
transformers==4.51.3
